requirements.txt
pandas
sqlalchemy
alpha_vantage
python-dotenv
